Split the upstream-marker tests out of test_async_support.py
The three pytester tests describe the `upstream` marker (#632), not async
test support, so they move to tests/test_upstream_marker.py together with
the httpx classifier test. What stays behind is only what this branch
claims: that an `async def test_` is collected and awaited, and that a
call through asyncio.to_thread agrees with the same call made directly.

The new module's docstring says why the marker needs a nested pytest run
at all: the hook rewrites a test's *outcome*, which the test itself
cannot observe -- a skipped test has already stopped and cannot assert
that it was skipped.
